Extract service layer from solution server
Split server.py (~1150 lines) into focused modules:
- settings.py: SolutionServerSettings
- resources.py: _SharedResources, KaiSolutionServerContext, with_db_recovery
- service.py: all business logic + new query/collection/bulk functions
- server.py: slimmed to ~200 lines (MCP tool wrappers only)

Added _get_kai_ctx() helper to eliminate repeated context extraction
boilerplate. Moved session_maker guard into with_db_recovery decorator.
Updated test imports accordingly.
